requirements.txt
pandas
numpy
SQLAlchemy
psycopg2-binary
catboost
scikit-learn
python-dotenv
msal
requests
streamlit
joblib
openpyxl
shap
matplotlib
seaborn